WIP: Migrate to Pusher Beams & add onboarding view
Because local notifications don't really suffice for this kind of app, I decided to switch to Pusher Beams.
This also implied lots of changes in the apps business logic, mainly the settings data structure and the
settings UI.

I also added a proper onboarding view to get started quickly.

This is still WIP, because the functionality to see incoming events listed in the app is gone (temporarily).
It has to be added back to the app.